Move keystore_geth.go to eth-node/bridge/geth
